// src/app/shared/model/ref-user-role-obj.model.ts
export class RefUserRole {
    RefUserRoleId: number;
    RefBizUnitId: number;
    RefJobTitleId: number;
    RefUserId: number;
    SpvId: number;
    RefRoleId: number;
    RefOfficeId: number;
    IsActive: boolean;
    RowVersion : string;

    constructor(){
    this.RefUserRoleId  = 0,
    this.RowVersion = ""
    }

}